Version 46 has been released
Update paths and version for new directory structure

Updated the `OutputPath` in the `.csproj` file to use the new `46` directory. Adjusted `HintPath` for multiple references to reflect the updated directory structure. Enhanced the `Microsoft.Dynamics.AX.Metadata.Extensions` reference with additional metadata and updated its path. Incremented the version in `source.extension.vsixmanifest` to `17.1.36.46` to align with the new release.